php dir and initial mysql table creation logic
the beauty of PHP comes from it's dovetail relationship and integration
with MySQL. The VERY FIRST thing to do is learn how to create and
inspect a table from within PHP and get it working w/ your MySQL
instance. This is that 1st step. :D Welcome to the desert of the real.